+++ app/src/sharedTest/net/chmielowski/randomchoice/AbstractTest.kt
@@ -6,14 +6,8 @@ import android.graphics.Bitmap
 import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.annotation.StringRes
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.asAndroidBitmap
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertTextEquals
 import androidx.compose.ui.test.captureToImage
@@ -24,11 +18,8 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.onRoot
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.test.platform.app.InstrumentationRegistry
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import net.chmielowski.randomchoice.core.Choice
 import net.chmielowski.randomchoice.core.MainExecutor
 import net.chmielowski.randomchoice.core.createStateStore
@@ -55,6 +46,8 @@ internal abstract class AbstractTest {
 
     protected open val prepopulateDatabase = PrepopulateDatabase {}
 
+    protected open val menuStrategy: DropdownMenuStrategy = DropdownMenuStrategy.Real()
+
     @Before
     fun setUp() {
         rule.setContent { Content() }
@@ -79,7 +72,7 @@ internal abstract class AbstractTest {
                     deleteDilemma = DeleteSavedDilemmaImpl(database, NonCancellableTask.fake),
                 )
             }),
-            menuStrategy = FakeDropdownMenuStrategy,
+            menuStrategy = menuStrategy,
         )
     }
 
@@ -264,39 +257,4 @@ internal fun fakeThemePreference(theme: Theme) = object : ThemePreference {
     override val current = theme
 
     override fun write(theme: Theme) {}
-}
-
-@Suppress("TestFunctionName")
-object FakeDropdownMenuStrategy : DropdownMenuStrategy {
-
-    @Composable
-    override fun Container(content: @Composable () -> Unit) = content()
-
-    @Composable
-    override fun Menu(
-        expanded: Boolean,
-        onDismissRequest: () -> Unit,
-        content: @Composable () -> Unit
-    ) {
-        if (expanded) {
-            content()
-        }
-    }
-
-    @Composable
-    override fun Item(
-        icon: ImageVector,
-        choice: Boolean?,
-        text: Int,
-        onClick: () -> Unit,
-        onDismiss: () -> Unit,
-    ) = Text(
-        stringResource(text),
-        modifier = Modifier.size(10.dp).clickable { onClick(); onDismiss() },
-        fontSize = 1.sp
-    )
-
-    @Composable
-    override fun Divider(modifier: Modifier) {
-    }
 }
